requirements.txt
plotly.express
pandas
seaborn==0.12.2
optbinning==0.17.3